resets circle properly on skip
Fully draws circle when a pomo ends or is skipped to ensure that future animations aren't messed up. Timer still does not work when time for pomos are low (< 1 min). It seems to work at higher times but more testing needs to be done to make sure there isn't actually a mistake in code.